+++ kalshi/client.py
@@ -9,6 +9,7 @@ from types import TracebackType
 from kalshi._base_client import SyncTransport
 from kalshi.auth import KalshiAuth
 from kalshi.config import DEMO_BASE_URL, DEMO_WS_URL, KalshiConfig
+from kalshi.resources.communications import CommunicationsResource
 from kalshi.resources.events import EventsResource
 from kalshi.resources.exchange import ExchangeResource
 from kalshi.resources.historical import HistoricalResource
@@ -18,6 +19,7 @@ from kalshi.resources.order_groups import OrderGroupsResource
 from kalshi.resources.orders import OrdersResource
 from kalshi.resources.portfolio import PortfolioResource
 from kalshi.resources.series import SeriesResource
+from kalshi.resources.subaccounts import SubaccountsResource
 
 
 class KalshiClient:
@@ -82,6 +84,7 @@ class KalshiClient:
 
         # Build transport and resources
         self._transport = SyncTransport(self._auth, self._config)
+        self.communications = CommunicationsResource(self._transport)
         self.events = EventsResource(self._transport)
         self.exchange = ExchangeResource(self._transport)
         self.historical = HistoricalResource(self._transport)
@@ -90,6 +93,7 @@ class KalshiClient:
         self.orders = OrdersResource(self._transport)
         self.portfolio = PortfolioResource(self._transport)
         self.series = SeriesResource(self._transport)
+        self.subaccounts = SubaccountsResource(self._transport)
         self.multivariate_collections = MultivariateCollectionsResource(self._transport)
 
     @property